nesting pong gameloop
