+++ srcs/containers/django/src/pong/static/js/pong_core.js
@@ -2,7 +2,7 @@
 
 function drawRect(x, y, width, height, color) {
 	ctx.fillStyle = color;
-    ctx.fillRect(x, y, width, height);
+	ctx.fillRect(x, y, width, height);
 }
 
 function drawPaddle(paddle) {
@@ -16,7 +16,7 @@ function drawBall(ball) {
 function drawNet() {
 	for (let i = 0; i < canvas.height; i += 29) {
 		drawRect(canvas.width / 2 - 7, i, 14, 14, monoColor);
-    }
+	}
 }
 
 // Update functions
@@ -24,11 +24,11 @@ function drawNet() {
 function updatePaddle(paddle) {
 	paddle.y += paddle.dy;
 	
-    if (paddle.y < 0) {
+	if (paddle.y < 0) {
 		paddle.y = 0;
 	}
 	
-    if (paddle.y + paddle.height > canvas.height) {
+	if (paddle.y + paddle.height > canvas.height) {
 		paddle.y = canvas.height - paddle.height;
 	}
 }
@@ -42,12 +42,12 @@ function abs(x) {
 
 function updateBall() {
 	ball.x += ball.dx;
-    ball.y += ball.dy;
+	ball.y += ball.dy;
 	
 	// Bounce off top and bottom
-    if (ball.y < 0 || ball.y + ball.height > canvas.height) {
+	if (ball.y < 0 || ball.y + ball.height > canvas.height) {
 		ball.dy *= -1; 
-    }
+	}
 	
 	// Bounce off paddles
 	if (ball.x > player1.x && ball.x <= player1.x + player1.width) {
@@ -89,10 +89,10 @@ function movePaddlesPlayer1() {
 	if (keys["w"]) {
 		player1.dy = -paddleSpeed;
 	}
-    else if (keys["s"]) {
+	else if (keys["s"]) {
 		player1.dy = paddleSpeed;
 	}
-    else {
+	else {
 		player1.dy = 0;
 	}
 }
@@ -116,4 +116,113 @@ function displayScore() {
 	ctx.fillStyle = monoColor;
 	ctx.fillText(player1Score, canvas.width / 2 - 80, 50);
 	ctx.fillText(player2Score, canvas.width / 2 + 52, 50);
+}
+
+function updateScore() {
+	if (ball.x < player1.x) {
+		player2Score += 1;
+		resetBall();
+	} else if (ball.x + ball.width > canvas.width) {
+		player1Score += 1;
+		resetBall();
+	}
+	if (player1Score === scoreToWin){
+		setTimeout(function() {
+			alert('Player 1 wins!');
+		} , 100);
+		return false;
+	}
+	else if (player2Score === scoreToWin){
+		setTimeout(function() {
+			alert('Player 2 wins!');
+		} , 100);
+		return false;
+	}
+	return true;
+}
+
+// Main loop
+function gameLoop(mode)
+{
+	// Clear the canvas
+	ctx.clearRect(0, 0, canvas.width, canvas.height);
+
+	// Draw net, paddles, and ball
+	drawNet();
+	drawPaddle(player1);
+	drawPaddle(player2);
+	drawBall(ball);
+	if (mode === 'tournament'){
+		displayScoreTournament();
+		if (!gameRunning){ 
+			player1Score = 0;
+			player2Score = 0;
+			if (player1Score === scoreToWin) {
+				scoreBoard[matchOrder[currentGameIndex - 1][0]] += 1;
+				console.log('number of victory player ' + matchOrder[currentGameIndex - 1][0] + ' :' + scoreBoard[matchOrder[currentGameIndex - 1][0]]);
+			} else {
+				scoreBoard[matchOrder[currentGameIndex - 1][1]] += 1;
+				console.log('number of victory player ' + matchOrder[currentGameIndex - 1][1] + ' :' + scoreBoard[matchOrder[currentGameIndex - 1][1]]);
+			}
+			nextGame();
+			return; 
+		}
+	}
+	else{
+		displayScore();
+		if (!gameRunning) return;
+	}
+
+	// Update game state
+	movePaddlesPlayer1();
+	if (mode === 'user-vs-user'){
+		movePaddlesPlayer2();
+	}
+	else if (mode === 'user-vs-computer'){
+		movePaddlesComputer();
+	}
+	updatePaddle(player1);
+	updatePaddle(player2);
+	updateBall();
+	if (mode === 'tournament'){
+		gameRunning = updateScoreTournament();
+	}
+	else{
+		gameRunning = updateScore();
+	}
+	requestAnimationFrame(gameLoop);
+}
+
+function startGameUserVsUser() {
+	const player2Name = document.getElementById('player2Name').value;
+	if (player2Name.trim() === '') {
+		alert('Please enter a valid name for Player 2');
+	} else {
+		loadPage('pong')
+		const player1Name = '{{ user.username|default:"Guest" }}';
+		startGame(player1Name, player2Name, 'user-vs-user');
+	}
+}
+
+function startGame(player1Name, player2Name, mode) {
+	document.getElementById('pongCanvas').style.display = 'block';
+    console.log(`Starting game: ${player1Name} vs ${player2Name}`);
+	player1Score = 0;
+	player2Score = 0;
+	gameRunning = true;
+	console.log(`Game mode: ${mode}`);
+	if (mode === 'tournament') {
+		gameLoop(mode);
+	}
+	else if (mode === 'user-vs-user') {
+		gameLoop(mode);
+	}
+	else if (mode === 'user-vs-computer') {
+		gameLoop(mode);
+	}
+}
+
+function endGame() {
+	var redirecturi = "/";
+	window.location.href = redirecturi;
 }
