+++ srcs/containers/django/src/pong/templates/main/base.html
@@ -25,8 +25,7 @@
     <script src="{% static 'js/page_loading.js' %}" defer ></script>
     <script src="{% static 'js/pong-conf.js' %}" defer ></script>
     <script src="{% static 'js/pong-tournament.js' %}" defer ></script>
-    <script src="{% static 'js/pong-user-vs-computer.js' %}" defer ></script>
-    <script src="{% static 'js/pong-user-vs-user.js' %}" defer ></script>
+    <script src="{% static 'js/pong-ai.js' %}" defer ></script>
     <script src="{% static 'js/tournament-init.js' %}" defer ></script>
 	  <script type="module" src="{% static 'js/3dapp.js' %}" defer></script>
   </body>
