+++ srcs/containers/django/src/pong/templates/main/two_player_local.html
@@ -15,7 +15,7 @@
   </div>
   <div class="d-grid gap-2 col-6 mx-auto">
     <button type="button" class="btn btn-outline-light"
-            href="#pong_two_player_local" onclick="startGameUserVsUser(); loadPage('pong_two_player_local')">Continue</button>
+            href="#pong_two_player_local" onclick="startGame('player 1', 'player 2', 'user-vs-computer'); loadPage('pong_two_player_local')">Continue</button>
   </div>
 </div>
 {% endblock %}
